+++ Package.swift
@@ -11,23 +11,17 @@ let package = Package(
     ],
     dependencies: [
         .package(url: "https://github.com/ionic-team/capacitor-swift-pm.git", branch: "main"),
-        .package(url: "https://github.com/urbanairship/airship-mobile-framework-proxy.git", from: "8.1.0")
+        .package(url: "https://github.com/urbanairship/airship-mobile-framework-proxy.git", from: "10.0.0")
     ],
     targets: [
          .target(
-            name: "UaCapacitorAirshipPlugin",
+            name: "UaCapacitorAirship",
             dependencies: [
                 .product(name: "Capacitor", package: "capacitor-swift-pm"),
                 .product(name: "Cordova", package: "capacitor-swift-pm"),
                 .product(name: "AirshipFrameworkProxy", package: "airship-mobile-framework-proxy")
             ],
             path: "ios/Plugin"
-        ),
-        .target(
-            name: "UaCapacitorAirship",
-            dependencies: [.target(name: "UaCapacitorAirshipPlugin")],
-            path: "ios/Bootloader",
-            publicHeadersPath: "Public"
         )
     ]
 )