+++ example/ios/App/ExampleNotificationServiceExtension/NotificationService.swift
@@ -1,0 +1,5 @@
+import AirshipServiceExtension
+import UserNotifications
+
+class NotificationService: UANotificationServiceExtension {
+}