+++ example/ios/App/ExampleWidgetExtension/ExampleAttributes.swift
@@ -1,0 +1,13 @@
+import ActivityKit
+import WidgetKit
+import SwiftUI
+
+struct ExampleAttributes: ActivityAttributes {
+    public struct ContentState: Codable, Hashable {
+        // Dynamic stateful properties about your activity go here!
+        var emoji: String
+    }
+
+    // Fixed non-changing properties about your activity go here!
+    var name: String
+}