+++ ios/Bootloader/Public/AirshipCapacitorBootstrap.h
@@ -1,8 +1,0 @@
-/* Copyright Airship and Contributors */
-
-@import Foundation;
-@import UIKit;
-
-@interface AirshipCapacitorBootstrap: NSObject
-
-@end